IV - F stat
It is true that the median is at 6.63, indicating that estimates may be noisier for more than half of the considered years. That said, the use of the lagged fas price as additional instrument ensures an overall strong predictive power of the first-stage equation, and consequently, accurate (i.e. less noisy) second-stage estimates.